+++ src/components/Timesheet.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import StyledTimesheet from '@styles/components/Timesheet.styled';
+
+export const Timesheet = (): JSX.Element =>
+  <StyledTimesheet>
+    Timesheet
+  </StyledTimesheet>;
+