+++ src/styles/components/Timesheet.styled.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const StyledTimesheet = styled.div`
+  height: 250px;
+  width: 250px;
+  background: #DDD;
+  border-radius: 5px;
+  padding: 10px;
+`;
+
+export default StyledTimesheet;